requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
scikit-learn
sqlmodel
apscheduler
schedule
openai
